# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Flora-Chloris Flowershop POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reservations/my:
    get:
      operationId: getMyReservations
      tags: [reservations]
      summary: Get reservations for the currently logged-in user
      responses:
        "200":
          description: List of current user's reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List all reservations (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation (online)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get reservation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    put:
      operationId: updateReservationStatus
      tags: [reservations]
      summary: Update reservation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationStatusRequest"
      responses:
        "200":
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        purchaseCost:
          type: number
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
          nullable: true
        stockQuantity:
          type: integer
        isAvailable:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
      required:
        - id
        - name
        - price
        - categoryId
        - stockQuantity
        - isAvailable
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        purchaseCost:
          type: number
          nullable: true
        categoryId:
          type: integer
        stockQuantity:
          type: integer
        imageUrl:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        isAvailable:
          type: boolean
      required:
        - name
        - price
        - categoryId
        - stockQuantity
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        purchaseCost:
          type: number
          nullable: true
        categoryId:
          type: integer
        stockQuantity:
          type: integer
        isAvailable:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required:
        - productId
        - quantity
        - unitPrice
        - subtotal
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        reservationNumber:
          type: string
        userId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        pickupDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        total:
          type: number
        notes:
          type: string
          nullable: true
        paymentMethod:
          type: string
          enum: [cod, gcash]
          nullable: true
        fulfillmentType:
          type: string
          enum: [pickup, delivery]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - reservationNumber
        - customerName
        - customerPhone
        - items
        - pickupDate
        - status
        - total
        - fulfillmentType
        - createdAt
    CreateReservationRequest:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
        pickupDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        paymentMethod:
          type: string
          enum: [cod, gcash]
          nullable: true
        fulfillmentType:
          type: string
          enum: [pickup, delivery]
      required:
        - customerName
        - customerPhone
        - items
        - pickupDate
    UpdateReservationStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
      required:
        - status
...