# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flora-Chloris Flowershop POS API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  purchaseCost?: number | null;
  categoryId: number;
  categoryName?: string | null;
  stockQuantity: number;
  isAvailable: boolean;
  imageUrl?: string | null;
  unit?: string | null;
}